src/main/java/practice/leecode/Math/CountPrimes_204.java
package practice.leecode.Math;

import java.util.Arrays;

public class CountPrimes_204 {

    public int countPrimes(int n) {
        boolean[] primes = new boolean[n];
        Arrays.fill(primes, true);
        int ans = 0;
        for (int i = 2; i < n; i++) {
            if (primes[i]) {
                ans += 1;
                if ((long) i * i < n) {
                    for (int j = i * i; j < n; j += i) {
                        primes[j] = false;
                    }
                }
            }
        }
        return ans;
    }

}
